Fixed crash, faction search and profile indexing.
Profiles were improperly looked up by the wrong
pagination in ReloadProfiles routine. Now we properly
use the cursor id to lookup the latest message in
the ChannelMessages() call.

A bad usage of strings.Trim was causing a bug in
faction search begining with S.

A wrong check when looking and cachig Guilds in-memory
could crash the bot once he joined a guild.