Band-aid fixes to restore some of the bot commands.
Due to some bad design the bot started to crawl
erroneously the website for the /server-info command.
This was observed by #22, since results were inconsistent.

This commit restores /stats and /arena as they are "cheap",
and disables /mods and /server-info as they are not working
anymore and require a lot of work.

Closes #29